src/main/test/java/com/aoc/y2022/ProblemDay2Test.java
package com.aoc.y2022;

import com.aoc.ProblemDayTest;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import static org.fest.assertions.Assertions.assertThat;

public class ProblemDay2Test implements ProblemDayTest {

    private ProblemDay2 problem;

    @BeforeEach
    public void before() {
        problem = new ProblemDay2();
    }

    @Test
    @Override
    public void testSolveA() {
        assertThat(problem.solveA()).isEqualTo(8933);
    }

    @Test
    @Override
    public void testSolveB() {
        assertThat(problem.solveB()).isEqualTo(11998);
    }

}
